requirements.txt
Flask==1.0.2
flask-restful==0.3.6
webargs
google-cloud-firestore
google-cloud-storage
flask-cors
pyyaml
google-cloud-logging
firebase-admin==2.13.0
sendgrid==5.6.0
google-api-python-client
google-auth
google-auth-httplib2
oauth2client==4.1.2
retrying==1.3.3
flask_restful_swagger
flasgger
